requirements.txt
python-telegram-bot>=20.0
redis>=4.5.0
python-dotenv>=1.0.0